LRCI-3823 Rearrange event classes to keep logic separate since GitHub events are so different
