+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/EventRestController.java
@@ -3,10 +3,8 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.event.controller;
+package com.liferay.jethr0.event;
 
-import com.liferay.jethr0.event.handler.EventHandler;
-import com.liferay.jethr0.event.handler.EventHandlerFactory;
 import com.liferay.jethr0.util.StringUtil;
 
 import org.apache.commons.logging.Log;
