+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/jenkins/BaseJenkinsEventHandler.java
@@ -3,10 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.event.handler;
+package com.liferay.jethr0.event.jenkins;
 
 import com.liferay.jethr0.bui1d.repository.BuildRunEntityRepository;
 import com.liferay.jethr0.bui1d.run.BuildRunEntity;
+import com.liferay.jethr0.event.BaseEventHandler;
+import com.liferay.jethr0.event.EventHandlerContext;
 import com.liferay.jethr0.jenkins.node.JenkinsNodeEntity;
 import com.liferay.jethr0.jenkins.repository.JenkinsServerEntityRepository;
 import com.liferay.jethr0.jenkins.server.JenkinsServerEntity;
@@ -21,7 +23,8 @@ import org.json.JSONObject;
 /**
  * @author Michael Hashimoto
  */
-public abstract class BaseJenkinsEventHandler extends BaseEventHandler {
+public abstract class BaseJenkinsEventHandler
+	extends BaseEventHandler implements JenkinsEventHandler {
 
 	protected BaseJenkinsEventHandler(
 		EventHandlerContext eventHandlerContext, JSONObject messageJSONObject) {
