+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/jenkins/BuildCompletedEventHandler.java
@@ -3,49 +3,61 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.event.handler;
+package com.liferay.jethr0.event.jenkins;
 
 import com.liferay.jethr0.bui1d.BuildEntity;
-import com.liferay.jethr0.bui1d.queue.BuildQueue;
 import com.liferay.jethr0.bui1d.repository.BuildEntityRepository;
 import com.liferay.jethr0.bui1d.repository.BuildRunEntityRepository;
 import com.liferay.jethr0.bui1d.run.BuildRunEntity;
+import com.liferay.jethr0.event.EventHandlerContext;
 import com.liferay.jethr0.job.JobEntity;
 import com.liferay.jethr0.job.repository.JobEntityRepository;
-
-import java.util.Date;
 
 import org.json.JSONObject;
 
 /**
  * @author Michael Hashimoto
  */
-public class BuildStartedEventHandler extends BaseJenkinsEventHandler {
+public class BuildCompletedEventHandler extends BaseJenkinsEventHandler {
+
+	public BuildCompletedEventHandler(
+		EventHandlerContext eventHandlerContext, JSONObject messageJSONObject) {
+
+		super(eventHandlerContext, messageJSONObject);
+	}
 
 	@Override
 	public String process() throws InvalidJSONException {
 		BuildRunEntity buildRunEntity = getBuildRun();
 
-		buildRunEntity.setJenkinsBuildURL(getJenkinsBuildURL());
-		buildRunEntity.setState(BuildRunEntity.State.RUNNING);
+		buildRunEntity.setDuration(getBuildDuration());
+		buildRunEntity.setResult(getBuildRunResult());
+		buildRunEntity.setState(BuildRunEntity.State.COMPLETED);
 
 		BuildEntity buildEntity = buildRunEntity.getBuildEntity();
 
-		buildEntity.setState(BuildEntity.State.RUNNING);
+		buildEntity.setState(BuildEntity.State.COMPLETED);
 
 		JobEntity jobEntity = buildEntity.getJobEntity();
 
-		if (jobEntity.getState() != JobEntity.State.RUNNING) {
-			jobEntity.setStartDate(new Date());
-			jobEntity.setState(JobEntity.State.RUNNING);
+		JobEntity.State jobState = JobEntity.State.COMPLETED;
+
+		for (BuildEntity jobBuildEntity : jobEntity.getBuildEntities()) {
+			BuildEntity.State buildState = jobBuildEntity.getState();
+
+			if (buildState != BuildEntity.State.COMPLETED) {
+				jobState = JobEntity.State.RUNNING;
+
+				break;
+			}
+		}
+
+		if (jobState == JobEntity.State.COMPLETED) {
+			jobEntity.setState(jobState);
 
 			JobEntityRepository jobEntityRepository = getJobEntityRepository();
 
 			jobEntityRepository.update(jobEntity);
-
-			BuildQueue buildQueue = getBuildQueue();
-
-			buildQueue.sort();
 		}
 
 		BuildEntityRepository buildEntityRepository = getBuildRepository();
@@ -57,15 +69,9 @@ public class BuildStartedEventHandler extends BaseJenkinsEventHandler {
 
 		buildRunEntityRepository.update(buildRunEntity);
 
-		updateJRPStatus(buildRunEntity, buildEntity, jobEntity, "running");
+		updateJRPStatus(buildRunEntity, buildEntity, jobEntity, "completed");
 
 		return buildRunEntity.toString();
-	}
-
-	protected BuildStartedEventHandler(
-		EventHandlerContext eventHandlerContext, JSONObject jsonObject) {
-
-		super(eventHandlerContext, jsonObject);
 	}
 
 }
