+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/jenkins/BuildStartedEventHandler.java
@@ -3,54 +3,56 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.event.handler;
+package com.liferay.jethr0.event.jenkins;
 
 import com.liferay.jethr0.bui1d.BuildEntity;
+import com.liferay.jethr0.bui1d.queue.BuildQueue;
 import com.liferay.jethr0.bui1d.repository.BuildEntityRepository;
 import com.liferay.jethr0.bui1d.repository.BuildRunEntityRepository;
 import com.liferay.jethr0.bui1d.run.BuildRunEntity;
+import com.liferay.jethr0.event.EventHandlerContext;
 import com.liferay.jethr0.job.JobEntity;
 import com.liferay.jethr0.job.repository.JobEntityRepository;
+
+import java.util.Date;
 
 import org.json.JSONObject;
 
 /**
  * @author Michael Hashimoto
  */
-public class BuildCompletedEventHandler extends BaseJenkinsEventHandler {
+public class BuildStartedEventHandler extends BaseJenkinsEventHandler {
+
+	public BuildStartedEventHandler(
+		EventHandlerContext eventHandlerContext, JSONObject jsonObject) {
+
+		super(eventHandlerContext, jsonObject);
+	}
 
 	@Override
 	public String process() throws InvalidJSONException {
 		BuildRunEntity buildRunEntity = getBuildRun();
 
-		buildRunEntity.setDuration(getBuildDuration());
-		buildRunEntity.setResult(getBuildRunResult());
-		buildRunEntity.setState(BuildRunEntity.State.COMPLETED);
+		buildRunEntity.setJenkinsBuildURL(getJenkinsBuildURL());
+		buildRunEntity.setState(BuildRunEntity.State.RUNNING);
 
 		BuildEntity buildEntity = buildRunEntity.getBuildEntity();
 
-		buildEntity.setState(BuildEntity.State.COMPLETED);
+		buildEntity.setState(BuildEntity.State.RUNNING);
 
 		JobEntity jobEntity = buildEntity.getJobEntity();
 
-		JobEntity.State jobState = JobEntity.State.COMPLETED;
-
-		for (BuildEntity jobBuildEntity : jobEntity.getBuildEntities()) {
-			BuildEntity.State buildState = jobBuildEntity.getState();
-
-			if (buildState != BuildEntity.State.COMPLETED) {
-				jobState = JobEntity.State.RUNNING;
-
-				break;
-			}
-		}
-
-		if (jobState == JobEntity.State.COMPLETED) {
-			jobEntity.setState(jobState);
+		if (jobEntity.getState() != JobEntity.State.RUNNING) {
+			jobEntity.setStartDate(new Date());
+			jobEntity.setState(JobEntity.State.RUNNING);
 
 			JobEntityRepository jobEntityRepository = getJobEntityRepository();
 
 			jobEntityRepository.update(jobEntity);
+
+			BuildQueue buildQueue = getBuildQueue();
+
+			buildQueue.sort();
 		}
 
 		BuildEntityRepository buildEntityRepository = getBuildRepository();
@@ -62,15 +64,9 @@ public class BuildCompletedEventHandler extends BaseJenkinsEventHandler {
 
 		buildRunEntityRepository.update(buildRunEntity);
 
-		updateJRPStatus(buildRunEntity, buildEntity, jobEntity, "completed");
+		updateJRPStatus(buildRunEntity, buildEntity, jobEntity, "running");
 
 		return buildRunEntity.toString();
-	}
-
-	protected BuildCompletedEventHandler(
-		EventHandlerContext eventHandlerContext, JSONObject messageJSONObject) {
-
-		super(eventHandlerContext, messageJSONObject);
 	}
 
 }
