+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/jenkins/JenkinsEventHandler.java
@@ -3,8 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.event.handler;
+package com.liferay.jethr0.event.jenkins;
 
+import com.liferay.jethr0.event.EventHandler;
 import com.liferay.jethr0.util.StringUtil;
 
 import java.util.Objects;
@@ -12,19 +13,16 @@ import java.util.Objects;
 /**
  * @author Michael Hashimoto
  */
-public interface EventHandler {
-
-	public String process() throws InvalidJSONException;
+public interface JenkinsEventHandler extends EventHandler {
 
 	public enum EventType {
 
 		BUILD_COMPLETED, BUILD_STARTED, COMPUTER_BUSY, COMPUTER_IDLE,
 		COMPUTER_OFFLINE, COMPUTER_ONLINE, COMPUTER_TEMPORARILY_OFFLINE,
-		COMPUTER_TEMPORARILY_ONLINE, CREATE_BUILD, CREATE_BUILD_RUN,
-		CREATE_JENKINS_COHORT, CREATE_JOB, QUEUE_ITEM_ENTER_BLOCKED,
+		COMPUTER_TEMPORARILY_ONLINE, QUEUE_ITEM_ENTER_BLOCKED,
 		QUEUE_ITEM_ENTER_BUILDABLE, QUEUE_ITEM_ENTER_WAITING,
 		QUEUE_ITEM_LEAVE_BLOCKED, QUEUE_ITEM_LEAVE_BUILDABLE,
-		QUEUE_ITEM_LEAVE_WAITING, QUEUE_ITEM_LEFT, QUEUE_JOB;
+		QUEUE_ITEM_LEAVE_WAITING, QUEUE_ITEM_LEFT;
 
 		public static EventType get(String eventTypeString) {
 			if (StringUtil.isNullOrEmpty(eventTypeString)) {
@@ -38,25 +36,6 @@ public interface EventHandler {
 			}
 
 			return null;
-		}
-
-	}
-
-	public class InvalidJSONException extends Exception {
-
-		public InvalidJSONException() {
-		}
-
-		public InvalidJSONException(String message) {
-			super(message);
-		}
-
-		public InvalidJSONException(String message, Throwable throwable) {
-			super(message, throwable);
-		}
-
-		public InvalidJSONException(Throwable throwable) {
-			super(throwable);
 		}
 
 	}
