+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/jrp/BaseJRPEventHandler.java
@@ -3,9 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.event.handler;
+package com.liferay.jethr0.event.jrp;
 
 import com.liferay.jethr0.bui1d.BuildEntity;
+import com.liferay.jethr0.event.BaseEventHandler;
+import com.liferay.jethr0.event.EventHandlerContext;
 import com.liferay.jethr0.job.JobEntity;
 import com.liferay.jethr0.job.repository.JobEntityRepository;
 import com.liferay.jethr0.util.StringUtil;
@@ -18,9 +20,10 @@ import org.json.JSONObject;
 /**
  * @author Michael Hashimoto
  */
-public abstract class BaseObjectEventHandler extends BaseEventHandler {
+public abstract class BaseJRPEventHandler
+	extends BaseEventHandler implements JRPEventHandler {
 
-	protected BaseObjectEventHandler(
+	protected BaseJRPEventHandler(
 		EventHandlerContext eventHandlerContext, JSONObject messageJSONObject) {
 
 		super(eventHandlerContext, messageJSONObject);
