+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/jrp/CreateBuildEventHandler.java
@@ -3,13 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.event.handler;
+package com.liferay.jethr0.event.jrp;
 
 import com.liferay.jethr0.bui1d.BuildEntity;
 import com.liferay.jethr0.bui1d.parameter.BuildParameterEntity;
 import com.liferay.jethr0.bui1d.queue.BuildQueue;
 import com.liferay.jethr0.bui1d.repository.BuildEntityRepository;
 import com.liferay.jethr0.bui1d.repository.BuildParameterEntityRepository;
+import com.liferay.jethr0.event.EventHandlerContext;
 import com.liferay.jethr0.jenkins.JenkinsQueue;
 import com.liferay.jethr0.job.JobEntity;
 import com.liferay.jethr0.job.repository.JobEntityRepository;
@@ -19,7 +20,13 @@ import org.json.JSONObject;
 /**
  * @author Michael Hashimoto
  */
-public class CreateBuildEventHandler extends BaseObjectEventHandler {
+public class CreateBuildEventHandler extends BaseJRPEventHandler {
+
+	public CreateBuildEventHandler(
+		EventHandlerContext eventHandlerContext, JSONObject messageJSONObject) {
+
+		super(eventHandlerContext, messageJSONObject);
+	}
 
 	@Override
 	public String process() throws InvalidJSONException {
@@ -69,12 +76,6 @@ public class CreateBuildEventHandler extends BaseObjectEventHandler {
 		updateJRPStatus(null, buildEntity, jobEntity, "queued");
 
 		return jobEntity.toString();
-	}
-
-	protected CreateBuildEventHandler(
-		EventHandlerContext eventHandlerContext, JSONObject messageJSONObject) {
-
-		super(eventHandlerContext, messageJSONObject);
 	}
 
 }
