+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/jrp/CreateJobEventHandler.java
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.event.handler;
+package com.liferay.jethr0.event.jrp;
 
 import com.liferay.jethr0.bui1d.BuildEntity;
 import com.liferay.jethr0.bui1d.repository.BuildEntityRepository;
 import com.liferay.jethr0.bui1d.repository.BuildParameterEntityRepository;
+import com.liferay.jethr0.event.EventHandlerContext;
 import com.liferay.jethr0.jenkins.repository.JenkinsCohortEntityRepository;
 import com.liferay.jethr0.job.JobEntity;
 import com.liferay.jethr0.job.repository.JobEntityRepository;
@@ -19,7 +20,13 @@ import org.json.JSONObject;
 /**
  * @author Michael Hashimoto
  */
-public class CreateJobEventHandler extends BaseObjectEventHandler {
+public class CreateJobEventHandler extends BaseJRPEventHandler {
+
+	public CreateJobEventHandler(
+		EventHandlerContext eventHandlerContext, JSONObject messageJSONObject) {
+
+		super(eventHandlerContext, messageJSONObject);
+	}
 
 	@Override
 	public String process() throws InvalidJSONException {
@@ -94,12 +101,6 @@ public class CreateJobEventHandler extends BaseObjectEventHandler {
 		jobEntityRepository.update(jobEntity);
 
 		return jobEntity.toString();
-	}
-
-	protected CreateJobEventHandler(
-		EventHandlerContext eventHandlerContext, JSONObject messageJSONObject) {
-
-		super(eventHandlerContext, messageJSONObject);
 	}
 
 }
