+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/jrp/QueueJobEventHandler.java
@@ -3,9 +3,10 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.jethr0.event.handler;
+package com.liferay.jethr0.event.jrp;
 
 import com.liferay.jethr0.bui1d.queue.BuildQueue;
+import com.liferay.jethr0.event.EventHandlerContext;
 import com.liferay.jethr0.jenkins.JenkinsQueue;
 import com.liferay.jethr0.job.JobEntity;
 import com.liferay.jethr0.job.repository.JobEntityRepository;
@@ -15,7 +16,13 @@ import org.json.JSONObject;
 /**
  * @author Michael Hashimoto
  */
-public class QueueJobEventHandler extends BaseObjectEventHandler {
+public class QueueJobEventHandler extends BaseJRPEventHandler {
+
+	public QueueJobEventHandler(
+		EventHandlerContext eventHandlerContext, JSONObject messageJSONObject) {
+
+		super(eventHandlerContext, messageJSONObject);
+	}
 
 	@Override
 	public String process() throws InvalidJSONException {
@@ -36,12 +43,6 @@ public class QueueJobEventHandler extends BaseObjectEventHandler {
 		jenkinsQueue.invoke();
 
 		return jobEntity.toString();
-	}
-
-	protected QueueJobEventHandler(
-		EventHandlerContext eventHandlerContext, JSONObject messageJSONObject) {
-
-		super(eventHandlerContext, messageJSONObject);
 	}
 
 }
